+++ tests/tests.py
@@ -379,6 +379,10 @@ def test_linked():
 	run('-a AAAAAAAAAA...TTTTTTTTTT', 'linked.fasta', 'linked.fasta')
 
 
+def test_linked_anchored():
+	run('-a AAAAAAAAAA...TTTTT$', 'linked-anchored.fasta', 'linked.fasta')
+
+
 def test_fasta():
 	run('-a TTAGACATATCTCCGTCG', 'small.fasta', 'small.fastq')